Standardize DXF group code fields to int16_t
Update entity fields and parsing/writing logic to use std::int16_t for DXF group codes that are 16-bit per the DXF spec. Refactor affected classes (e.g., Spline, Polyline, Vertex, Hatch, Dimension, Leader, Viewport) and their methods for correct type usage. Improves DXF compliance and reduces risk of data misinterpretation.

+++ Core Libraries/EoDxfLib/EoDxfDrawing.cpp
@@ -419,7 +419,7 @@ bool EoDxfWrite::WriteSpline(EoDxfSpline* spline) {
   }
 
   // Fit points (11, 21, 31)
-  for (std::int32_t i = 0; i < spline->m_numberOfFitPoints; ++i) {
+  for (std::int16_t i = 0; i < spline->m_numberOfFitPoints; ++i) {
     const auto* fitPoint = spline->m_fitPoints[static_cast<size_t>(i)];
     m_writer->WriteDouble(11, fitPoint->x);
     m_writer->WriteDouble(21, fitPoint->y);
